+++ access-checkout/src/test/java/com/worldpay/access/checkout/validation/filters/LengthFilterFactoryTest.kt
@@ -19,7 +19,7 @@ class LengthFilterFactoryTest {
 
     @Test
     fun `should get pan length filter`() {
-        val filter : PanLengthFilter = lengthFilterFactory.getPanLengthFilter()
+        val filter : PanLengthFilter = lengthFilterFactory.getPanLengthFilter(false)
         assertNotNull(filter)
     }